+++ docs/script/search_index.js
@@ -126,6 +126,18 @@ window.esdocSearchIndex = [
     "class"
   ],
   [
+    "formiojs/src/components/datamap/datamap.form.js~datamap",
+    "function/index.html#static-function-DataMap",
+    "<span>DataMap</span> <span class=\"search-result-import-path\">formiojs/src/components/datamap/DataMap.form.js</span>",
+    "function"
+  ],
+  [
+    "formiojs/src/components/datamap/datamap.js~datamapcomponent",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html",
+    "<span>DataMapComponent</span> <span class=\"search-result-import-path\">formiojs/src/components/datamap/DataMap.js</span>",
+    "class"
+  ],
+  [
     "formiojs/src/components/datetime/datetime.form.js~datetime",
     "function/index.html#static-function-DateTime",
     "<span>DateTime</span> <span class=\"search-result-import-path\">formiojs/src/components/datetime/DateTime.form.js</span>",
@@ -588,12 +600,6 @@ window.esdocSearchIndex = [
     "function"
   ],
   [
-    "formiojs/src/utils/utils.js~convertflatpickrtoformat",
-    "function/index.html#static-function-convertFlatpickrToFormat",
-    "<span>convertFlatpickrToFormat</span> <span class=\"search-result-import-path\">formiojs/src/utils/utils.js</span>",
-    "function"
-  ],
-  [
     "formiojs/src/utils/utils.js~convertformattoflatpickr",
     "function/index.html#static-function-convertFormatToFlatpickr",
     "<span>convertFormatToFlatpickr</span> <span class=\"search-result-import-path\">formiojs/src/utils/utils.js</span>",
@@ -669,6 +675,12 @@ window.esdocSearchIndex = [
     "formiojs/src/utils/utils.js~formatdate",
     "function/index.html#static-function-formatDate",
     "<span>formatDate</span> <span class=\"search-result-import-path\">formiojs/src/utils/utils.js</span>",
+    "function"
+  ],
+  [
+    "formiojs/src/utils/utils.js~formatoffset",
+    "function/index.html#static-function-formatOffset",
+    "<span>formatOffset</span> <span class=\"search-result-import-path\">formiojs/src/utils/utils.js</span>",
     "function"
   ],
   [
@@ -774,6 +786,18 @@ window.esdocSearchIndex = [
     "function"
   ],
   [
+    "formiojs/src/utils/utils.js~iteratekey",
+    "function/index.html#static-function-iterateKey",
+    "<span>iterateKey</span> <span class=\"search-result-import-path\">formiojs/src/utils/utils.js</span>",
+    "function"
+  ],
+  [
+    "formiojs/src/utils/utils.js~loadzones",
+    "function/index.html#static-function-loadZones",
+    "<span>loadZones</span> <span class=\"search-result-import-path\">formiojs/src/utils/utils.js</span>",
+    "function"
+  ],
+  [
     "formiojs/src/utils/jsonlogic/operators.js~lodashoperators",
     "variable/index.html#static-variable-lodashOperators",
     "<span>lodashOperators</span> <span class=\"search-result-import-path\">formiojs/src/utils/jsonlogic/operators.js</span>",
@@ -813,6 +837,18 @@ window.esdocSearchIndex = [
     "formiojs/src/utils/utils.js~setactionproperty",
     "function/index.html#static-function-setActionProperty",
     "<span>setActionProperty</span> <span class=\"search-result-import-path\">formiojs/src/utils/utils.js</span>",
+    "function"
+  ],
+  [
+    "formiojs/src/utils/utils.js~timezonetext",
+    "function/index.html#static-function-timezoneText",
+    "<span>timezoneText</span> <span class=\"search-result-import-path\">formiojs/src/utils/utils.js</span>",
+    "function"
+  ],
+  [
+    "formiojs/src/utils/utils.js~uniquekey",
+    "function/index.html#static-function-uniqueKey",
+    "<span>uniqueKey</span> <span class=\"search-result-import-path\">formiojs/src/utils/utils.js</span>",
     "function"
   ],
   [
@@ -1789,13 +1825,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/formio.js~formio.token",
-    "class/src/Formio.js~Formio.html#static-set-token",
+    "class/src/Formio.js~Formio.html#static-get-token",
     "src/Formio.js~Formio.token",
     "member"
   ],
   [
     "src/formio.js~formio.token",
-    "class/src/Formio.js~Formio.html#static-get-token",
+    "class/src/Formio.js~Formio.html#static-set-token",
     "src/Formio.js~Formio.token",
     "member"
   ],
@@ -2208,12 +2244,6 @@ window.esdocSearchIndex = [
     "method"
   ],
   [
-    "src/webform.js~webform#hastimezone",
-    "class/src/Webform.js~Webform.html#instance-get-hasTimezone",
-    "src/Webform.js~Webform#hasTimezone",
-    "member"
-  ],
-  [
     "src/webform.js~webform#isbuilt",
     "class/src/Webform.js~Webform.html#instance-member-isBuilt",
     "src/Webform.js~Webform#isBuilt",
@@ -2245,13 +2275,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/webform.js~webform#loading",
-    "class/src/Webform.js~Webform.html#instance-get-loading",
+    "class/src/Webform.js~Webform.html#instance-set-loading",
     "src/Webform.js~Webform#loading",
     "member"
   ],
   [
     "src/webform.js~webform#loading",
-    "class/src/Webform.js~Webform.html#instance-set-loading",
+    "class/src/Webform.js~Webform.html#instance-get-loading",
     "src/Webform.js~Webform#loading",
     "member"
   ],
@@ -2424,12 +2454,6 @@ window.esdocSearchIndex = [
     "member"
   ],
   [
-    "src/webform.js~webform#submissionoffset",
-    "class/src/Webform.js~Webform.html#instance-get-submissionOffset",
-    "src/Webform.js~Webform#submissionOffset",
-    "member"
-  ],
-  [
     "src/webform.js~webform#submissionready",
     "class/src/Webform.js~Webform.html#instance-member-submissionReady",
     "src/Webform.js~Webform#submissionReady",
@@ -2445,12 +2469,6 @@ window.esdocSearchIndex = [
     "src/webform.js~webform#submissionreadyresolve",
     "class/src/Webform.js~Webform.html#instance-member-submissionReadyResolve",
     "src/Webform.js~Webform#submissionReadyResolve",
-    "member"
-  ],
-  [
-    "src/webform.js~webform#submissiontimezone",
-    "class/src/Webform.js~Webform.html#instance-get-submissionTimezone",
-    "src/Webform.js~Webform#submissionTimezone",
     "member"
   ],
   [
@@ -3013,13 +3031,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/wizardbuilder.js~wizardbuilder#currentpage",
-    "class/src/WizardBuilder.js~WizardBuilder.html#instance-set-currentPage",
+    "class/src/WizardBuilder.js~WizardBuilder.html#instance-get-currentPage",
     "src/WizardBuilder.js~WizardBuilder#currentPage",
     "member"
   ],
   [
     "src/wizardbuilder.js~wizardbuilder#currentpage",
-    "class/src/WizardBuilder.js~WizardBuilder.html#instance-get-currentPage",
+    "class/src/WizardBuilder.js~WizardBuilder.html#instance-set-currentPage",
     "src/WizardBuilder.js~WizardBuilder#currentPage",
     "member"
   ],
@@ -3697,13 +3715,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/components/base/base.js~basecomponent#disabled",
-    "class/src/components/base/Base.js~BaseComponent.html#instance-get-disabled",
+    "class/src/components/base/Base.js~BaseComponent.html#instance-set-disabled",
     "src/components/base/Base.js~BaseComponent#disabled",
     "member"
   ],
   [
     "src/components/base/base.js~basecomponent#disabled",
-    "class/src/components/base/Base.js~BaseComponent.html#instance-set-disabled",
+    "class/src/components/base/Base.js~BaseComponent.html#instance-get-disabled",
     "src/components/base/Base.js~BaseComponent#disabled",
     "member"
   ],
@@ -4723,13 +4741,13 @@ window.esdocSearchIndex = [
   ],
   [
     "src/components/checkbox/checkbox.js~checkboxcomponent#datavalue",
-    "class/src/components/checkbox/Checkbox.js~CheckBoxComponent.html#instance-set-dataValue",
+    "class/src/components/checkbox/Checkbox.js~CheckBoxComponent.html#instance-get-dataValue",
     "src/components/checkbox/Checkbox.js~CheckBoxComponent#dataValue",
     "member"
   ],
   [
     "src/components/checkbox/checkbox.js~checkboxcomponent#datavalue",
-    "class/src/components/checkbox/Checkbox.js~CheckBoxComponent.html#instance-get-dataValue",
+    "class/src/components/checkbox/Checkbox.js~CheckBoxComponent.html#instance-set-dataValue",
     "src/components/checkbox/Checkbox.js~CheckBoxComponent#dataValue",
     "member"
   ],
@@ -5574,6 +5592,198 @@ window.esdocSearchIndex = [
     "file"
   ],
   [
+    "src/components/datamap/datamap.form.js",
+    "file/src/components/datamap/DataMap.form.js.html",
+    "src/components/datamap/DataMap.form.js",
+    "file"
+  ],
+  [
+    "src/components/datamap/datamap.js",
+    "file/src/components/datamap/DataMap.js.html",
+    "src/components/datamap/DataMap.js",
+    "file"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#addkeybutton",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-addKeyButton",
+    "src/components/datamap/DataMap.js~DataMapComponent#addKeyButton",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#addrow",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-addRow",
+    "src/components/datamap/DataMap.js~DataMapComponent#addRow",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#build",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-build",
+    "src/components/datamap/DataMap.js~DataMapComponent#build",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#buildrow",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-buildRow",
+    "src/components/datamap/DataMap.js~DataMapComponent#buildRow",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#buildrows",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-buildRows",
+    "src/components/datamap/DataMap.js~DataMapComponent#buildRows",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#componentcomponents",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-get-componentComponents",
+    "src/components/datamap/DataMap.js~DataMapComponent#componentComponents",
+    "member"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#constructor",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-constructor-constructor",
+    "src/components/datamap/DataMap.js~DataMapComponent#constructor",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#createheader",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-createHeader",
+    "src/components/datamap/DataMap.js~DataMapComponent#createHeader",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#createvaluecomponent",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-createValueComponent",
+    "src/components/datamap/DataMap.js~DataMapComponent#createValueComponent",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#datavalue",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-member-dataValue",
+    "src/components/datamap/DataMap.js~DataMapComponent#dataValue",
+    "member"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#defaultschema",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-get-defaultSchema",
+    "src/components/datamap/DataMap.js~DataMapComponent#defaultSchema",
+    "member"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#emptyvalue",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-get-emptyValue",
+    "src/components/datamap/DataMap.js~DataMapComponent#emptyValue",
+    "member"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#getvalue",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-getValue",
+    "src/components/datamap/DataMap.js~DataMapComponent#getValue",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#hasaddbutton",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-hasAddButton",
+    "src/components/datamap/DataMap.js~DataMapComponent#hasAddButton",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#haschanged",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-hasChanged",
+    "src/components/datamap/DataMap.js~DataMapComponent#hasChanged",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#hasremovebuttons",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-hasRemoveButtons",
+    "src/components/datamap/DataMap.js~DataMapComponent#hasRemoveButtons",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#removekeybutton",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-removeKeyButton",
+    "src/components/datamap/DataMap.js~DataMapComponent#removeKeyButton",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#removerow",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-removeRow",
+    "src/components/datamap/DataMap.js~DataMapComponent#removeRow",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#rows",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-member-rows",
+    "src/components/datamap/DataMap.js~DataMapComponent#rows",
+    "member"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#schema",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-get-schema",
+    "src/components/datamap/DataMap.js~DataMapComponent#schema",
+    "member"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#setvalue",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-setValue",
+    "src/components/datamap/DataMap.js~DataMapComponent#setValue",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#tableelement",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-member-tableElement",
+    "src/components/datamap/DataMap.js~DataMapComponent#tableElement",
+    "member"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#type",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-member-type",
+    "src/components/datamap/DataMap.js~DataMapComponent#type",
+    "member"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent#updatevalue",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#instance-method-updateValue",
+    "src/components/datamap/DataMap.js~DataMapComponent#updateValue",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent.builderinfo",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#static-get-builderInfo",
+    "src/components/datamap/DataMap.js~DataMapComponent.builderInfo",
+    "member"
+  ],
+  [
+    "src/components/datamap/datamap.js~datamapcomponent.schema",
+    "class/src/components/datamap/DataMap.js~DataMapComponent.html#static-method-schema",
+    "src/components/datamap/DataMap.js~DataMapComponent.schema",
+    "method"
+  ],
+  [
+    "src/components/datamap/datamap.spec.js",
+    "file/src/components/datamap/DataMap.spec.js.html",
+    "src/components/datamap/DataMap.spec.js",
+    "file"
+  ],
+  [
+    "src/components/datamap/editform/datamap.edit.display.js",
+    "file/src/components/datamap/editForm/DataMap.edit.display.js.html",
+    "src/components/datamap/editForm/DataMap.edit.display.js",
+    "file"
+  ],
+  [
+    "src/components/datamap/fixtures/comp1.js",
+    "file/src/components/datamap/fixtures/comp1.js.html",
+    "src/components/datamap/fixtures/comp1.js",
+    "file"
+  ],
+  [
+    "src/components/datamap/fixtures/index.js",
+    "file/src/components/datamap/fixtures/index.js.html",
+    "src/components/datamap/fixtures/index.js",
+    "file"
+  ],
+  [
     "src/components/datetime/datetime.form.js",
     "file/src/components/datetime/DateTime.form.js.html",
     "src/components/datetime/DateTime.form.js",
@@ -5614,6 +5824,12 @@ window.esdocSearchIndex = [
     "class/src/components/datetime/DateTime.js~DateTimeComponent.html#instance-method-createWrapper",
     "src/components/datetime/DateTime.js~DateTimeComponent#createWrapper",
     "method"
+  ],
+  [
+    "src/components/datetime/datetime.js~datetimecomponent#dateformat",
+    "class/src/components/datetime/DateTime.js~DateTimeComponent.html#instance-get-dateFormat",
+    "src/components/datetime/DateTime.js~DateTimeComponent#dateFormat",
+    "member"
   ],
   [
     "src/components/datetime/datetime.js~datetimecomponent#datetimeformat",
@@ -5715,6 +5931,12 @@ window.esdocSearchIndex = [
     "src/components/datetime/datetime.js~datetimecomponent#validationvalue",
     "class/src/components/datetime/DateTime.js~DateTimeComponent.html#instance-get-validationValue",
     "src/components/datetime/DateTime.js~DateTimeComponent#validationValue",
+    "member"
+  ],
+  [
+    "src/components/datetime/datetime.js~datetimecomponent.default_format",
+    "class/src/components/datetime/DateTime.js~DateTimeComponent.html#static-get-DEFAULT_FORMAT",
+    "src/components/datetime/DateTime.js~DateTimeComponent.DEFAULT_FORMAT",
     "member"
   ],
   [