+++ gulpfile.js
@@ -77,7 +77,12 @@ gulp.task('scripts-formio', build('Formio.js', 'formio.js'));
 gulp.task('scripts-utils', build('utils/utils.js', 'formio.utils.js'));
 gulp.task('scripts-full', build('index.js', 'formio.full.js'));
 gulp.task('scripts-form', build('formio.form.js', 'formio.form.js'));
-gulp.task('scripts-embed', build('formio.embed.js', 'formio.embed.js'));
+gulp.task('formio.embed.min.js', () => buildProd('formio.embed.js', 'formio.embed.min.js'));
+gulp.task('formio.embed.js', () =>
+  gulp.src('./dist/formio.embed.min.js')
+    .pipe(plugins.rename('formio.embed.js'))
+    .pipe(gulp.dest('dist')));
+gulp.task('scripts-embed', gulpsync.sync([['formio.embed.min.js'], 'formio.embed.js']));
 gulp.task('scripts-contrib', build('contrib/index.js', 'formio.contrib.js'));
 
 // ESLint