+++ src/components/textarea/TextArea.js
@@ -134,7 +134,7 @@ export default class TextAreaComponent extends TextFieldComponent {
       }
 
       return quill;
-    });
+    }).catch(err => console.warn(err));
 
     return this.input;
   }